Move map replay table to a separate component
